Converted output headers to use CR_LF instead of platform dependent println().
NOTE: I'm concerned about potential problems with double-byte character encodings in headers here. The spec kind of implies (by using the word OCTET), that it's in 8859-1, but until now it was using platform encoding.

Please let me know of any breakages, or even explicit points in the spec where it says what the header encoding should be